# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HRPF API",
    "version": "0.4.0",
    "description": "Backend foundations, managed public content and atomic publication controls. Fixed NGO page copy is bundled in the frontend and needs no publication approval. Operational member/case review workflows remain later milestones. Public reads are no-store and withdrawals take effect immediately. Browser clients use the frontend same-origin /api rewrite and credentials. All mutating auth/admin requests require the exact configured Origin and X-CSRF-Token. Production cookies use Secure, HttpOnly, SameSite=Lax, Path=/, no Domain; development names omit __Host-."
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Administrator login",
        "security": [
          {
            "csrfHeader": [],
            "csrfCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string",
                              "format": "email",
                              "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                            },
                            "role": {
                              "type": "string",
                              "enum": [
                                "super_admin",
                                "admin",
                                "editor",
                                "case_manager"
                              ]
                            }
                          },
                          "required": [
                            "id",
                            "name",
                            "email",
                            "role"
                          ]
                        },
                        "csrfToken": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "user",
                        "csrfToken"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "summary": "Queue reset without disclosing account existence",
        "security": [
          {
            "csrfHeader": [],
            "csrfCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "const": "accepted"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "Consume reset token and revoke all user sessions",
        "security": [
          {
            "csrfHeader": [],
            "csrfCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "const": "password_reset"
                        }
                      },
                      "required": [
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/forms/session": {
      "post": {
        "summary": "Issue a 15-minute purpose-bound form ticket after Turnstile verification",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormSessionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "ticket": {
                          "type": "string"
                        },
                        "purpose": {
                          "type": "string",
                          "enum": [
                            "complaint",
                            "membership",
                            "contact"
                          ]
                        },
                        "expiresIn": {
                          "type": "number",
                          "const": 900
                        }
                      },
                      "required": [
                        "ticket",
                        "purpose",
                        "expiresIn"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/complaints": {
      "post": {
        "summary": "Persist validated form and email outbox atomically",
        "description": "Use the original ticket and UUID submissionKey for retries within the ticket lifetime. Same key with different data returns 409. Received confirms storage, not email delivery or approval. Membership requires an enabled policy.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "reference": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "const": "received"
                        }
                      },
                      "required": [
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/membership-applications": {
      "post": {
        "summary": "Persist validated form and email outbox atomically",
        "description": "Use the original ticket and UUID submissionKey for retries within the ticket lifetime. Same key with different data returns 409. Received confirms storage, not email delivery or approval. Membership requires an enabled policy.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "reference": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "const": "received"
                        }
                      },
                      "required": [
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/contact-messages": {
      "post": {
        "summary": "Persist validated form and email outbox atomically",
        "description": "Use the original ticket and UUID submissionKey for retries within the ticket lifetime. Same key with different data returns 409. Received confirms storage, not email delivery or approval. Membership requires an enabled policy.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "reference": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "const": "received"
                        }
                      },
                      "required": [
                        "status"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/form-uploads": {
      "post": {
        "summary": "Scan and stage one restricted file; JPG/PNG/WebP 5 MB or PDF 10 MB",
        "security": [
          {
            "formTicketHeader": []
          }
        ],
        "parameters": [
          {
            "name": "purpose",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "complaint",
                "membership"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "assetId": {
                          "type": "string"
                        },
                        "format": {
                          "type": "string"
                        },
                        "bytes": {
                          "type": "number"
                        },
                        "scanStatus": {
                          "type": "string",
                          "const": "clean"
                        }
                      },
                      "required": [
                        "assetId",
                        "format",
                        "bytes",
                        "scanStatus"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplaintInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "fatherName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9 ()-]{7,30}$"
          },
          "province": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "district": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "ticket": {
            "type": "string",
            "minLength": 43,
            "maxLength": 100
          },
          "submissionKey": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "consent": {
            "type": "boolean",
            "const": true
          },
          "consentVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "cnic": {
            "type": "string",
            "pattern": "^(?:\\d{13}|\\d{5}-\\d{7}-\\d)$"
          },
          "cnicImageId": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "complaintDocumentId": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000
          },
          "priorProceedings": {
            "type": "boolean"
          },
          "priorProceedingsDetails": {
            "type": "string",
            "maxLength": 5000
          },
          "decisionDocumentIds": {
            "default": [],
            "maxItems": 3,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-f0-9]{24}$"
            }
          },
          "attachmentIds": {
            "default": [],
            "maxItems": 3,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-f0-9]{24}$"
            }
          }
        },
        "required": [
          "name",
          "fatherName",
          "email",
          "phone",
          "province",
          "district",
          "address",
          "ticket",
          "submissionKey",
          "consent",
          "consentVersion",
          "cnic",
          "cnicImageId",
          "complaintDocumentId",
          "category",
          "description",
          "priorProceedings"
        ],
        "additionalProperties": false
      },
      "MembershipInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "fatherName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9 ()-]{7,30}$"
          },
          "province": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "district": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "ticket": {
            "type": "string",
            "minLength": 43,
            "maxLength": 100
          },
          "submissionKey": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "consent": {
            "type": "boolean",
            "const": true
          },
          "consentVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "membershipType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "paymentReference": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "paymentProofId": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          }
        },
        "required": [
          "name",
          "fatherName",
          "email",
          "phone",
          "province",
          "district",
          "address",
          "ticket",
          "submissionKey",
          "consent",
          "consentVersion",
          "membershipType",
          "paymentReference",
          "paymentProofId"
        ],
        "additionalProperties": false
      },
      "ContactInput": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "string",
            "minLength": 43,
            "maxLength": 100
          },
          "submissionKey": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "consent": {
            "type": "boolean",
            "const": true
          },
          "consentVersion": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9 ()-]{7,30}$"
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 5000
          }
        },
        "required": [
          "ticket",
          "submissionKey",
          "consent",
          "consentVersion",
          "name",
          "email",
          "subject",
          "message"
        ],
        "additionalProperties": false
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          }
        },
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false
      },
      "FormSessionInput": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "complaint",
              "membership",
              "contact"
            ]
          },
          "botToken": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2048
          }
        },
        "required": [
          "purpose",
          "botToken"
        ],
        "additionalProperties": false
      },
      "ForgotInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
      "ResetInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 43,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 200
          }
        },
        "required": [
          "token",
          "password"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-access"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-refresh"
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-csrf"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "formTicketHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Form-Ticket"
      }
    }
  }
}